+++ pydocx/openxml/wordprocessing/__init__.py
@@ -1,0 +1,12 @@
+# coding: utf-8
+from pydocx.openxml.wordprocessing.paragraph_properties import ParagraphProperties  # noqa
+from pydocx.openxml.wordprocessing.run_properties import RunProperties  # noqa
+from pydocx.openxml.wordprocessing.style import Style
+from pydocx.openxml.wordprocessing.styles import Styles
+
+__all__ = [
+    'ParagraphProperties',
+    'RunProperties',
+    'Style',
+    'Styles',
+]